Reuse the household wizard base logic code
To follow DRY.
